added chain logic, missing the chain itself
Full implementation of chain logic. Added the ChainConstraint script, which attaches to the player and controls math. Also slightly modified the player movement, to adapt functions from the ChainConstraint script.

Player is limited to a elliptical area when chain is enabled. Ellipse size can be modified using the properties on the script.

To add:
- physics-based chain sprite that follows the player
- (maybe) movement/velocity correction, depending on if it feels good during playtest